% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_delta}
\alias{compute_delta}
\title{Definition of progression deltas for different tests.}
\usage{
compute_delta(baseline, outcome = "edss")
}
\arguments{
\item{baseline}{Outcome value at baseline.}

\item{outcome}{One of: \cr
'edss' (Extended Disability Status Scale ) [default]; \cr
'nhptD' (Nine-Hole Peg Test, dominant hand) \cr
'nhptND' (Nine-Hole Peg Test, non-dominant hand) \cr
't25fw' (Timed 25-Foot Walk) \cr
'sdmt' (Symbol Digit Modalities Test).}
}
\value{
Minimum delta corresponding to valid change from the provided baseline value.
}
\description{
Definition of progression deltas for different tests.
}
\examples{
compute_delta(4.5)
compute_delta(55, outcome='sdmt')
}
