% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msprog_methods.R
\name{criteria_text}
\alias{criteria_text}
\title{Textual description of criteria used to obtain object.}
\usage{
criteria_text(object)
}
\arguments{
\item{object}{An object for which a textual description of criteria is desired.}
}
\value{
The form of the value returned by \code{criteria_text} depends on the class of its argument.
See the documentation of the particular methods for details of what is produced by that method.
}
\description{
Textual description of criteria used to obtain object.
}
