% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msprog_methods.R
\name{event_count.MSprogOutput}
\alias{event_count.MSprogOutput}
\title{Event count for disability progression results.}
\usage{
\method{event_count}{MSprogOutput}(object)
}
\arguments{
\item{object}{An object of class \code{'MSprogOutput'} (result of a call to \code{\link[=MSprog]{MSprog()}}).}
}
\value{
A \code{data.frame} object containing the sequence of events for each subject,
as well as the event count separated by event type
(improvement, progression, RAW, PIRA, undefined progression).
}
\description{
\code{event_count} method for class \code{'MSprogOutput'}.
}
\examples{
# EDSS progression
output <- MSprog(toydata_visits, 'id', 'EDSS', 'date', 'edss',
    relapse=toydata_relapses, conf_weeks=12, conf_tol_days=30,
    event='multiple', baseline='roving', verbose=2)
print(event_count(output)) # event sequence and count for each subject
}
