% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msprog_methods.R
\name{event_count}
\alias{event_count}
\title{Event count from object.}
\usage{
event_count(object)
}
\arguments{
\item{object}{An object for which an event count is desired.}
}
\value{
The form of the value returned by \code{event_count} depends on the class of its argument.
See the documentation of the particular methods for details of what is produced by that method.
}
\description{
Event count from object.
}
