% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relapse_indep_from_bounds.R
\name{relapse_indep_from_bounds}
\alias{relapse_indep_from_bounds}
\title{Define relapse-free intervals for PIRA definition.}
\usage{
relapse_indep_from_bounds(
  b0 = 0,
  b1 = 0,
  e0 = 90,
  e1 = 30,
  c0 = 90,
  c1 = 30,
  use_end_dates = F
)
}
\arguments{
\item{b0}{Days before baseline (\verb{>=0}).}

\item{b1}{Days after baseline (\verb{>=0}), or \code{NULL}.}

\item{e0}{Days before event (\verb{>=0}), or \code{NULL}.}

\item{e1}{Days after event (\verb{>=0}), or \code{NULL}.}

\item{c0}{Days before confirmation (\verb{>=0}), or \code{NULL}.}

\item{c1}{Days after confirmation (\verb{>=0}).}

\item{use_end_dates}{If \code{TRUE}, only the right bounds (\code{e1}, \code{c1}) are used,
as the right bounds will be defined by the onset-to-end interval of each relapse.
This option is only relevant when relapse end dates are available.}
}
\value{
A named list to be given as argument \code{relapse_indep} to function \code{\link[=MSprog]{MSprog()}}
}
\description{
\code{relapse_indep_from_bounds()} organises the given interval bounds around baseline, event, and confirmation
into a named list to be given as argument \code{relapse_indep} to function \code{\link[=MSprog]{MSprog()}}.
}
\details{
If the right end is \code{NULL}, the interval is assumed to extend up to the left end of the next interval.
If the left end is \code{NULL}, the interval is assumed to extend up to the right end of the previous interval.
}
\examples{
# No relapses between baseline and confirmation (high-specificity definition from [1]):
relapse_indep <- relapse_indep_from_bounds(0,NULL,NULL,NULL,NULL,0)
# No relapses within event-90dd->event+30dd
# and within confirmation-90dd->confirmation+30dd [1]:
relapse_indep <- relapse_indep_from_bounds(0,0,90,30,90,30)
# No relapses within baseline->event+30dd and within confirmation+-30dd [2]:
relapse_indep <- relapse_indep_from_bounds(0,NULL,NULL,30,30,30)
}
\references{
[1] Müller J, Cagol A, Lorscheider J, Tsagkas C, Benkert P, Yaldizli Ö, et al.
Harmonizing definitions for progression independent of relapse activity in multiple sclerosis: A systematic review.
JAMA Neurol. 2023;80:1232–45. \cr\cr
[2] Kappos L, Wolinsky JS, Giovannoni G, Arnold DL, Wang Q, Bernasconi C, et al.
Contribution of relapse-independent progression vs relapse-associated worsening to overall confirmed disability
accumulation in typical relapsing multiple sclerosis in a pooled analysis of 2 randomized clinical trials.
JAMA Neurol. 2020;77:1132–40.
}
