% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msprog_methods.R
\name{results.MSprogOutput}
\alias{results.MSprogOutput}
\title{Extended disability progression results.}
\usage{
\method{results}{MSprogOutput}(object)
}
\arguments{
\item{object}{An object of class \code{'MSprogOutput'} (result of a call to \code{\link[=MSprog]{MSprog()}}).}
}
\value{
A \code{data.frame} object containing an extended report of all events
detected by function \code{\link[=MSprog]{MSprog()}} for each subject.
}
\description{
\code{results} method for class \code{'MSprogOutput'}.
}
\examples{
# EDSS progression
output <- MSprog(toydata_visits, 'id', 'EDSS', 'date', 'edss',
    relapse=toydata_relapses, conf_weeks=12, conf_tol_days=30,
    event='multiple', baseline='roving', verbose=2)
print(results(output)) # extended event info for each subject
}
