% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msprog_methods.R
\name{results}
\alias{results}
\title{Extended result report from object.}
\usage{
results(object)
}
\arguments{
\item{object}{An object for which an extended result report is desired.}
}
\value{
The form of the value returned by \code{results} depends on the class of its argument.
See the documentation of the particular methods for details of what is produced by that method.
}
\description{
Extended result report from object.
}
\keyword{internal}
