% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata_relapses}
\alias{toydata_relapses}
\title{Synthetic Relapse Data}
\format{
An object of class \code{data.frame}, with columns:
\describe{
\item{id}{Subject IDs.}
\item{date}{Relapse onset dates.}
}
}
\usage{
data(toydata_relapses)
}
\description{
Artificially generated relapse onset dates for some example patients in \link{toydata_visits}
to illustrate the use of the package.
}
\examples{

data(toydata_relapses)
head(toydata_relapses)

}
\references{
This data set was artificially created for the \code{msprog} package.
}
\keyword{datasets}
