% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata_visits}
\alias{toydata_visits}
\title{Synthetic Longitudinal EDSS and SDMT Data}
\format{
An object of class \code{data.frame}, with columns:
\describe{
\item{id}{Subject IDs.}
\item{date}{Visit dates.}
\item{EDSS}{Synthetic EDSS scores (values between 0 and 10).}
\item{SDMT}{Synthetic SDMT scores (values between 0 and 110).}
}
}
\usage{
data(toydata_visits)
}
\description{
Artificially generated toy data set including Extended Disability Status Scale (EDSS)
and Symbol Digit Modalities Test (SDMT) scores in a small cohort of example patients
to illustrate the use of the package.
}
\examples{

head(toydata_visits)

}
\references{
This data set was artificially created for the \code{msprog} package.
}
\keyword{datasets}
