% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_testing.R
\name{is_event}
\alias{is_event}
\title{Compare value to reference.}
\usage{
is_event(
  x,
  baseline,
  type,
  outcome = "edss",
  worsening = NULL,
  delta_fun = NULL,
  sub_threshold = FALSE
)
}
\arguments{
\item{x}{Outcome value to test.}

\item{baseline}{Outcome value at baseline.}

\item{type}{One of: \cr
\itemize{
\item{\code{'prog'}}{ (progression);}
\item{\code{'impr'}}{ (improvement);}
\item{\code{'change'}}{ (any valid change).}
}}

\item{outcome}{One of: \cr
\itemize{
\item{\code{'edss'}}{ (Extended Disability Status Scale, default);}
\item{\code{'nhpt'}}{ (Nine-Hole Peg Test);}
\item{\code{'t25fw'}}{ (Timed 25-Foot Walk);}
\item{\code{'sdmt'}}{ (Symbol Digit Modalities Test);}
\item{\code{NULL}}{ (only accepted when specifying the direction of worsening).}
}}

\item{worsening}{The direction of worsening (\code{'increase'} if higher values correspond to worse disease course, \code{'decrease'} otherwise).
This argument is only used when \code{outcome} is set to \code{NULL}.
If \code{outcome} is specified, \code{worsening} is automatically set to \code{'increase'} for EDSS, NHPT, T25FW,
and to \code{'decrease'} for SDMT.}

\item{delta_fun}{Custom function specifying the minimum shift corresponding
to a valid change from the provided baseline value. If none is specified (default),
\code{\link[=compute_delta]{compute_delta()}} for the specified outcome is used.}

\item{sub_threshold}{If \code{TRUE}, any confirmed progression, or improvement, or change in the outcome measure is valid,
regardless of \code{delta_fun}.}
}
\value{
A boolean value specifying if a valid event was found.
}
\description{
\code{is_event()} checks if an outcome value determines a valid progression,
or improvement, or change, from a given reference value.
}
\examples{
is_event(x=4.5, baseline=4, type='prog', outcome='edss')
is_event(x=50, baseline=57, type='prog', outcome='sdmt')
}
