% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSprog.R
\name{MSprog}
\alias{MSprog}
\title{Compute multiple sclerosis progression from longitudinal data.}
\usage{
MSprog(
  data,
  subj_col,
  value_col,
  date_col,
  subjects = NULL,
  relapse = NULL,
  rsubj_col = NULL,
  rdate_col = NULL,
  outcome = "edss",
  delta_fun = NULL,
  conf_months = c(6, 12),
  conf_tol_days = 45,
  conf_left = FALSE,
  require_sust_months = 0,
  rel_infl = 30,
  event = "multiple",
  baseline = "roving",
  sub_threshold = FALSE,
  relapse_rebl = FALSE,
  min_value = 0,
  prog_last_visit = FALSE,
  include_dates = FALSE,
  include_value = FALSE,
  verbose = 2
)
}
\arguments{
\item{data}{data frame containing longitudinal data, including: subject ID, outcome value, date of visit.}

\item{subj_col}{Name of data column with subject ID.}

\item{value_col}{Name of data column with outcome value.}

\item{date_col}{Name of data column with date of visit.}

\item{subjects}{(optional) Subset of subjects.}

\item{relapse}{(optional) data frame containing longitudinal data, including: subject ID and relapse date.}

\item{rsubj_col}{Name of subject ID column for relapse data, if different from outcome data.}

\item{rdate_col}{Name of subject ID column for relapse data, if different from outcome data.}

\item{outcome}{One of: \cr
'edss' (Extended Disability Status Scale ) [default]; \cr
'nhptD' (Nine-Hole Peg Test, dominant hand) \cr
'nhptND' (Nine-Hole Peg Test, non-dominant hand) \cr
't25fw' (Timed 25-Foot Walk) \cr
'sdmt' (Symbol Digit Modalities Test).}

\item{delta_fun}{(optional) Custom function specifying the minimum delta corresponding
to a valid change from the provided baseline value.}

\item{conf_months}{Period before confirmation (months).}

\item{conf_tol_days}{Tolerance window for confirmation visit (days); can be an integer (same tolerance on left and right)
or list-like of length 2 (different tolerance on left and right).}

\item{conf_left}{If TRUE, confirmation window is unbounded on the right.}

\item{require_sust_months}{Minimum number of months from confirmation for which a change must be sustained to be retained as an event.}

\item{rel_infl}{Influence of last relapse (days).}

\item{event}{One of: \cr
'first' (only the very first event - improvement or progression); \cr
'firsteach' (first improvement and first progression); \cr
'firstprog' (first progression); \cr
'firstprogtype' (first progression of each kind - PIRA, RAW, undefined); \cr
'multiple' (all events) [default]. \cr}

\item{baseline}{One of: \cr
'fixed' (first outcome value out of relapse influence); \cr
'roving' (updated after each event to last confirmed outcome value out of relapse influence) [default].}

\item{sub_threshold}{If TRUE, include confirmed sub-threshold events for roving baseline.}

\item{relapse_rebl}{If TRUE, search for PIRA events again with post-relapse re-baseline.}

\item{min_value}{Only consider progressions events where the outcome is >= value.}

\item{prog_last_visit}{If TRUE, include progressions occurring at last visit (i.e. with no confirmation).}

\item{include_dates}{If TRUE, report dates of events.}

\item{include_value}{If TRUE, report value of outcome at event.}

\item{verbose}{One of: \cr
0 (print no info); \cr
1 (print concise info); \cr
2 (print extended info) [default].}
}
\value{
Two data.frame objects: \cr
- summary of event sequence detected for each subject; \cr
- extended info on each event for all subjects. \cr
}
\description{
`MSprog` detects and characterises the progression (or improvement) events of an outcome measure
(EDSS, NHPT, T25FW, or SDMT) for one or more subjects, based on repeated assessments
through time and on the dates of acute episodes.
Several qualitative and quantitative options are given as arguments that can be set
by the user and reported as a complement to the results to ensure reproducibility.
}
\details{
The events are detected sequentially by scanning the outcome values in chronological order.
Progression events are classified as relapse-associated or relapse-independent based on their relative timing
with respect to the relapses. Specifically, relapse-associated worsening (RAW) events
are defined as confirmed progression events occurring within the influence of a relapse,
while progression independent of relapse activity (PIRA) is established when the progression
event occurs out of relapse influence, and with no relapses between baseline and confirmation.
}
\examples{
data(toydata_visits)
data(toydata_relapses)
# EDSS progression
output <- MSprog(toydata_visits, 'id', 'EDSS', 'date', relapse=toydata_relapses,
    outcome='edss', conf_months=3, conf_tol_days=30, rel_infl=30,
    event='multiple', baseline='roving', verbose=2)
summary_EDSS = output[[1]] # summary of event sequence for each subject
results_EDSS = output[[2]] # extended info on each event for all subjects
# SDMT progression
output <- MSprog(toydata_visits, 'id', 'SDMT', 'date', relapse=toydata_relapses,
    outcome='sdmt', conf_months=3, conf_tol_days=30, rel_infl=30,
    event='multiple', baseline='roving', verbose=2)
summary_SDMT = output[[1]] # summary of event sequence for each subject
results_SDMT = output[[2]] # extended info on each event for all subjects
}
