% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSprog.R
\name{MSprog}
\alias{MSprog}
\title{Compute multiple sclerosis progression from longitudinal data.}
\usage{
MSprog(
  data,
  subj_col,
  value_col,
  date_col,
  subjects = NULL,
  relapse = NULL,
  rsubj_col = NULL,
  rdate_col = NULL,
  outcome = "edss",
  delta_fun = NULL,
  conf_months = 3,
  conf_tol_days = 30,
  conf_left = FALSE,
  require_sust_months = 0,
  rel_infl = 30,
  event = "firstprog",
  baseline = "fixed",
  pira_def = 0,
  sub_threshold = FALSE,
  relapse_rebl = FALSE,
  min_value = 0,
  prog_last_visit = FALSE,
  include_dates = FALSE,
  include_value = FALSE,
  include_stable = TRUE,
  verbose = 1
)
}
\arguments{
\item{data}{\code{data.frame} containing longitudinal data, including: subject ID, outcome value, date of visit.}

\item{subj_col}{Name of data column with subject ID.}

\item{value_col}{Name of data column with outcome value.}

\item{date_col}{Name of data column with date of visit.}

\item{subjects}{Subset of subjects (list of IDs) [default is \code{NULL}].}

\item{relapse}{\code{data.frame} containing longitudinal data, including: subject ID and relapse date [default is \code{NULL}].}

\item{rsubj_col}{Name of subject ID column for relapse data, if different from outcome data [default is \code{NULL}].}

\item{rdate_col}{Name of subject ID column for relapse data, if different from outcome data [default is \code{NULL}].}

\item{outcome}{One of:
\itemize{
\item{\code{'edss'}}{ (Extended Disability Status Scale ) [default];}
 \item{\code{'nhpt'}}{ (Nine-Hole Peg Test);}
 \item{\code{'nhptD'}}{ (Nine-Hole Peg Test, dominant hand);}
 \item{\code{'nhptND'}}{ (Nine-Hole Peg Test, non-dominant hand);}
 \item{\code{'t25fw'}}{ (Timed 25-Foot Walk);}
 \item{\code{'sdmt'}}{ (Symbol Digit Modalities Test).}
 }}

\item{delta_fun}{Custom function specifying the minimum delta corresponding
to a valid change from the provided baseline value. If none is specified [default], \code{compute_delta} for the specified outcome is used.}

\item{conf_months}{Period before confirmation (months) [default is 3].}

\item{conf_tol_days}{Tolerance window for confirmation visit (days); can be an integer (same tolerance on left and right)
or list-like of length 2 (different tolerance on left and right) [default is 45].}

\item{conf_left}{If \code{TRUE}, confirmation window is unbounded on the right [default is \code{FALSE}].}

\item{require_sust_months}{Minimum number of months from confirmation for which a change must be sustained to be retained as an event [default is 0].}

\item{rel_infl}{Influence of last relapse (days) [default is 30].}

\item{event}{One of:
\itemize{
\item{\code{'firstprog'}}{ (first progression) [default];}
\item{\code{'first'}}{ (only the very first event - improvement or progression);}
\item{\code{'firsteach'}}{ (first improvement and first progression);}
\item{\code{'firstprogtype'}}{ (first progression of each kind - PIRA, RAW, undefined);}
\item{\code{'multiple'}}{ (all events).}
}}

\item{baseline}{One of:
\itemize{
\item{\code{'fixed'}}{ (first outcome value out of relapse influence) [default];}
\item{\code{'roving'}}{ (updated after each event to last confirmed outcome value out of relapse influence).}
}}

\item{pira_def}{One of:
\itemize{
 \item{0}{ (no relapses between baseline and confirmation,
as in [Cagol et al, JAMA Neurol 2022, doi:10.1001/jamaneurol.2022.1025]);}
\item{1}{ (no relapses in [baseline, progression+\code{rel_infl}] and in [confirmation-\code{rel_infl}, confirmation+\code{rel_infl}],
as in [Kappos et al, JAMA Neurol 2020, doi:10.1001/jamaneurol.2020.1568]).}
}}

\item{sub_threshold}{If \code{TRUE}, include confirmed sub-threshold events for roving baseline [default is \code{FALSE}].}

\item{relapse_rebl}{If \code{TRUE}, search for PIRA events again with post-relapse re-baseline [default is \code{FALSE}].}

\item{min_value}{Only consider progressions events where the outcome is >= value [default is 0].}

\item{prog_last_visit}{If \code{TRUE}, include progressions occurring at last visit (i.e. with no confirmation) [default is \code{FALSE}].}

\item{include_dates}{If \code{TRUE}, report dates of events [default is \code{FALSE}].}

\item{include_value}{If \code{TRUE}, report value of outcome at event [default is \code{FALSE}].}

\item{include_stable}{If \code{TRUE}, include subjects with no events in extended output \code{data.frame} with time2event = total follow up [default is \code{TRUE}].}

\item{verbose}{One of:
\itemize{
 \item{0}{ (print no info);}
 \item{1}{ (print concise info) [default];}
 \item{2}{ (print extended info).}
 }}
}
\value{
Two \code{data.frame} objects:
\itemize{
\item{summary of event sequence detected for each subject;}
\item{extended info on each event for all subjects.}
}
}
\description{
\code{MSprog} detects and characterises the progression (or improvement) events of an outcome measure
(EDSS, NHPT, T25FW, or SDMT) for one or more subjects, based on repeated assessments
through time and on the dates of acute episodes.
Several qualitative and quantitative options are given as arguments that can be set
by the user and reported as a complement to the results to ensure reproducibility.
}
\details{
The events are detected sequentially by scanning the outcome values in chronological order.
Progression events are classified as relapse-associated or relapse-independent based on their relative timing
with respect to the relapses. Specifically, relapse-associated worsening (RAW) events
are defined as confirmed progression events occurring within the influence of a relapse,
while progression independent of relapse activity (PIRA) is established when the progression
event occurs out of relapse influence, and with no relapses between baseline and confirmation.
}
\examples{
data(toydata_visits)
data(toydata_relapses)
# EDSS progression
output <- MSprog(toydata_visits, 'id', 'EDSS', 'date', relapse=toydata_relapses,
    outcome='edss', conf_months=3, conf_tol_days=30, rel_infl=30,
    event='multiple', baseline='roving', verbose=2)
summary_EDSS = output[[1]] # summary of event sequence for each subject
results_EDSS = output[[2]] # extended info on each event for all subjects
# SDMT progression
output <- MSprog(toydata_visits, 'id', 'SDMT', 'date', relapse=toydata_relapses,
    outcome='sdmt', conf_months=3, conf_tol_days=30, rel_infl=30,
    event='multiple', baseline='roving', verbose=2)
summary_SDMT <- output[[1]] # summary of event sequence for each subject
results_SDMT <- output[[2]] # extended info on each event for all subjects
}
