% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msprog_methods.R
\name{print.MSprogOutput}
\alias{print.MSprogOutput}
\title{Textual description of criteria used to compute disability progression.}
\usage{
\method{print}{MSprogOutput}(x, ...)
}
\arguments{
\item{x}{An object of class \code{'MSprogOutput'} (result of a call to \code{\link[=MSprog]{MSprog()}}).}

\item{...}{Optional arguments for \code{print} methods. They are ignored in this function.}
}
\description{
\code{print} method for class \code{'MSprogOutput'}.
}
\details{
The method prints out a short paragraph describing the set of criteria used to
obtain the output.
}
\examples{
# EDSS progression
output <- MSprog(toydata_visits, 'id', 'EDSS', 'date', 'edss',
    relapse=toydata_relapses, conf_days=7*12, conf_tol_days=30,
    event='multiple', baseline='roving', verbose=2)
print(output) # textual description of parameters used to obtain output
}
