% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata_relapses}
\alias{toydata_relapses}
\title{Synthetic Relapse Data}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{id}{Subject ID}
 \item{date}{The relapse date}
}
}
\usage{
data(toydata_relapses)
}
\description{
Artificially generated relapse dates for 3 example patients to illustrate the use of the package.
}
\examples{

data(toydata_relapses)
head(toydata_relapses)

}
\references{
This data set was artificially created for the msprog package.
}
\keyword{datasets}
