% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata_visits}
\alias{toydata_visits}
\title{Synthetic Longitudinal EDSS and SDMT Data}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{id}{Subject ID}
 \item{date}{The visit date}
 \item{EDSS}{A value between 0 and 10}
 \item{SDMT}{A value between 0 and 110}
}
}
\usage{
data(toydata_visits)
}
\description{
Small, artificially generated toy data set providing Extended Disability Status Scale (EDSS)
and Symbol Digit Modalities Test (SDMT) information on 3 example patients
to illustrate the use of the package.
}
\examples{

data(toydata_visits)
head(toydata_visits)

}
\references{
This data set was artificially created for the msprog package.
}
\keyword{datasets}
