% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_milestone.R
\name{value_milestone}
\alias{value_milestone}
\title{Time to disability milestone.}
\usage{
value_milestone(
  data,
  milestone,
  subj_col,
  value_col,
  date_col,
  outcome,
  worsening = NULL,
  relapse = NULL,
  rsubj_col = NULL,
  rdate_col = NULL,
  validconf_col = NULL,
  conf_days = 24 * 7,
  conf_tol_days = c(7, 2 * 365.25),
  require_sust_days = 0,
  relapse_to_event = 0,
  relapse_to_conf = 30,
  impute_last_visit = 0,
  verbose = 0
)
}
\arguments{
\item{data}{a \code{data.frame} containing longitudinal data, including: subject ID, outcome value, date of visit.}

\item{milestone}{Disability milestone (outcome value to check data against).}

\item{subj_col}{Name of data column with subject ID.}

\item{value_col}{Name of data column with outcome value.}

\item{date_col}{Name of data column with date of visit.}

\item{outcome}{Specifies the outcome type. Must be one of the following:
\itemize{
\item{\code{'edss'} (Expanded Disability Status Scale);}
\item{\code{'nhpt'} (Nine-Hole Peg Test);}
\item{\code{'t25fw'} (Timed 25-Foot Walk);}
\item{\code{'sdmt'} (Symbol Digit Modalities Test);}
\item{\code{NULL} (only accepted when specifying argument \code{worsening})}
}}

\item{worsening}{The direction of worsening (\code{'increase'} if higher values correspond to worse disease course, \code{'decrease'} otherwise).\if{html}{\out{<br />}}
The given value is only used when \code{outcome} is set to \code{NULL}. Otherwise, \code{worsening} is automatically set to
\code{'increase'} if \code{outcome} is set to \code{'edss'}, \code{'nhpt'}, \code{'t25fw'},
and to \code{'decrease'} if \code{outcome} is set to \code{'sdmt'}.}

\item{relapse}{\code{data.frame} containing longitudinal data, including: subject ID and relapse date.}

\item{rsubj_col}{Name of subject column for relapse data, if different from outcome data.}

\item{rdate_col}{Name of date column for relapse data, if different from outcome data.}

\item{validconf_col}{Name of data column specifying which visits can (\code{T}) or cannot (\code{F}) be used as confirmation visits.
The input data does not necessarily have to include such a column.
If \code{validconf_col=NULL}, all visits are potentially used as confirmation visits.}

\item{conf_days}{Period before confirmation (days).}

\item{conf_tol_days}{Tolerance window for confirmation visit (days); can be an integer (same tolerance on left and right)
or list-like of length 2 (different tolerance on left and right).
The right end of the interval can be set to \code{Inf} (confirmation window unbounded on the right
-- e.g., "confirmed over 12 \emph{or more} weeks").}

\item{require_sust_days}{Minimum number of days over which the milestone must be sustained
(i.e., confirmed at \emph{all} visits occurring in the specified period).
If the milestone is sustained for the remainder of the follow-up period, it is considered reached regardless of follow-up duration.
If \code{require_sust_days=Inf}, values are retained only when sustained for the remainder of the follow-up period.}

\item{relapse_to_event}{Minimum distance from a relapse (days) for the milestone to be considered reached.
Can be an integer (minimum distance from \emph{last} relapse) or list-like of length 2
(minimum distance from \emph{last} relapse, minimum distance from \emph{next} relapse).
Note that setting the distance to zero means retaining the event regardless of surrounding relapses.}

\item{relapse_to_conf}{Minimum distance from a relapse (days) for a visit to be a valid confirmation visit.
Can be an integer (minimum distance from \emph{last} relapse) or list-like of length 2
(minimum distance from \emph{last} relapse, minimum distance from \emph{next} relapse).
Note that setting the distance to zero means using any visit for confirmation regardless of surrounding relapses.}

\item{impute_last_visit}{Imputation probability when the milestone is reached at the last visit (i.e. with no confirmation).
Unconfirmed values exceeding the milestone at the last visit are never imputed if \code{impute_last_visit=0};
they are always imputed if \code{impute_last_visit=1};
they are imputed with probability \code{p}, \verb{0<p<1}, if \code{impute_last_visit=p}.
If a value \code{N>1} is passed, unconfirmed values exceeding the milestone are imputed only if occurring within \code{N} days of follow-up
(e.g., in case of early discontinuation).}

\item{verbose, }{One of:
\itemize{
\item{0}{ (print no info);}
\item{1}{ (print concise info, default);}
\item{2}{ (print extended info).}
}}
}
\value{
A \code{data.frame} containing the following columns:
\itemize{
\item{\code{date_col}: }{the date of first reaching or exceeding the milestone (or last date of follow-up if milestone is not reached);}
\item{\code{value_col}: }{the first value  reaching orexceeding the milestone, if present, otherwise no value;}
\item{\code{'time2event'}: }{the time taken to reach or exceed the milestone (or total follow-up length if milestone is not reached);}
\item{\code{'observed'}: }{whether the milestone was reached (1) or not (0).}
}
}
\description{
\code{value_milestone()} scans the visits in chronological order to detect the first
outcome value exceeding a specified disability milestone (e.g., EDSS>=6), \emph{with confirmation}.
Note: "exceeding" means either value>=milestone or value<=milestone, depending on the
outcome measure (see arguments \code{outcome} and \code{worsening}).
}
\details{
An event is only retained if \strong{confirmed}, i.e., if all values \emph{up to} the
confirmation visit exceed the milestone.
Valid time windows for confirmation visits are determined by arguments
\code{conf_days}, \code{conf_tol_days}, \code{relapse_to_conf}.
}
