% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_milestone.R
\name{value_milestone}
\alias{value_milestone}
\title{Time to disability milestone.}
\usage{
value_milestone(
  data,
  milestone,
  value_col,
  date_col,
  subj_col,
  outcome,
  worsening = NULL,
  relapse = NULL,
  rsubj_col = NULL,
  rdate_col = NULL,
  conf_weeks = 24,
  conf_tol_days = 30,
  conf_unbounded_right = FALSE,
  relapse_to_event = 0,
  relapse_to_conf = 30,
  verbose = 0
)
}
\arguments{
\item{data}{a \code{data.frame} containing longitudinal data containing subject ID, outcome value, date of visit.}

\item{milestone}{Disability milestone (outcome value to check data against).}

\item{value_col}{Name of data column with outcome value.}

\item{date_col}{Name of data column with date of visit.}

\item{subj_col}{Name of data column with subject ID.}

\item{outcome}{Specifies the outcome type. Must be one of the following:
\itemize{
\item{\code{'edss'}}{ (Expanded Disability Status Scale);}
\item{\code{'nhpt'}}{ (Nine-Hole Peg Test);}
\item{\code{'t25fw'}}{ (Timed 25-Foot Walk);}
\item{\code{'sdmt'}}{ (Symbol Digit Modalities Test);}
\item{\code{NULL}}{ (only accepted when specifying argument \code{worsening})}
}}

\item{worsening}{The direction of worsening (\code{'increase'} if higher values correspond to worse disease course, \code{'decrease'} otherwise).\if{html}{\out{<br />}}
This argument is only used when \code{outcome} is set to \code{NULL}. Otherwise, \code{worsening} is automatically set to
\code{'increase'} if \code{outcome} is set to \code{'edss'}, \code{'nhpt'}, \code{'t25fw'},
and to \code{'decrease'} if \code{outcome} is set to \code{'sdmt'}.}

\item{relapse}{\code{data.frame} containing longitudinal data, including: subject ID and relapse date.}

\item{rsubj_col}{Name of subject column for relapse data, if different from outcome data.}

\item{rdate_col}{Name of date column for relapse data, if different from outcome data.}

\item{conf_weeks}{Period before confirmation (weeks).}

\item{conf_tol_days}{Tolerance window for confirmation visit (days).}

\item{conf_unbounded_right}{If \code{TRUE}, confirmation window is unbounded on the right.}

\item{relapse_to_event}{Minimum distance from a relapse (days) for an outcome value to be valid.}

\item{relapse_to_conf}{Minimum distance from a relapse (days) for a valid confirmation visit.}

\item{verbose, }{One of:
\itemize{
\item{0}{ (print no info);}
\item{1}{ (print concise info, default);}
\item{2}{ (print extended info).}
}}
}
\value{
A \code{data.frame} containing the following columns:
\itemize{
\item{\code{date_col}: }{the date of first reaching a value >= milestone (or last date of follow-up if milestone is not reached);}
\item{\code{value_col}: }{the first value >= milestone, if present, otherwise no value is reported;}
\item{\code{'time2event'}: }{the time to reach a value >= milestone (or total follow-up length if milestone is not reached);}
\item{\code{'observed'}: }{whether the milestone was reached (1) or not (0).}
}
}
\description{
\code{value_milestone()} scans the visits in chronological order to detect the first
outcome value exceeding a specified disability milestone (e.g., EDSS>=6), \emph{with confirmation}.
Note: "exceeding" means either value>=milestone or value<=milestone, depending on the
outcome measure (see arguments \code{outcome} and \code{worsening}).
}
\details{
An event is only retained if \strong{confirmed}, i.e., if all values \emph{up to} the
confirmation visit exceed the milestone.
Valid time windows for confirmation visits are determined by arguments
\code{conf_weeks}, \code{conf_tol_days}, \code{conf_unbounded_right}, \code{relapse_to_conf}.
}
